requirements.txt
Django==2.0.3
djangorestframework==3.7.7
requests==2.18.4
django-bower==5.0.4
mysqlclient==1.3.12
django-filter
django-bootstrap-form
python-gedcom
pika
Celery==4.1.0
